pyproject.toml
[tool.poetry]
name = "cheaters_bowl"
version = "0.1.0"
description = ""
authors = ["Andrew <amao1@umd.edu>"]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.61.1"
SQLAlchemy = "^1.3.20"
PyJWT = "^1.7.1"
wikipedia = "^1.4.0"
uvloop = "^0.14.0"
uvicorn = "^0.12.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
Wikipedia-API = "^0.5.4"
Unidecode = "^1.1.1"
pandas = "^1.1.4"
python-jose = {extras = ["cryptography"], version = "^3.2.0"}


[tool.poetry.dev-dependencies]
pylint = "^2.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
